#1294 Add to support ZK Web debugging using IntelliJ https://github.com/adempiere/adempiere/issues/1294 - Update the Scala and SBT configuration
(cherry picked from commit 7043d6eddea11c4196e7a80b107729ee3d7b9a55)
