+++ org.adempiere.test/project/build.properties
