+++ org.adempiere.test/project/plugins.sbt
@@ -1,5 +1,4 @@
 import _root_.sbt._
-import _root_.sbt._
 
 addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.13.0")
 addSbtPlugin("com.earldouglas" % "xsbt-web-plugin" % "2.1.0")
