+++ project/plugins.sbt
@@ -2,4 +2,5 @@ import _root_.sbt._
 import _root_.sbt._
 
 addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.13.0")
-addSbtPlugin("com.earldouglas" % "xsbt-web-plugin" % "2.1.0")
+addSbtPlugin("com.earldouglas" % "xsbt-web-plugin" % "4.0.1")
+addSbtPlugin("com.github.mpeltonen" % "sbt-idea" % "1.6.0")
